added sections

+++ PaginaToros/Shared/Models/EstadoHistoricoRe.cs
@@ -5,17 +5,17 @@ namespace PaginaToros.Shared.Models;
 
 public partial class EstadoHistoricoRe
 {
-    public double? Ea1 { get; set; } = 0;
+    public int? Ea1 { get; set; } = 0;
 
-    public double? Ea2 { get; set; } = 0;
+    public int? Ea2 { get; set; } = 0;
 
-    public double? Ea3 { get; set; } = 0;
+    public int? Ea3 { get; set; } = 0;
 
-    public double? Ea4 { get; set; } = 0;
+    public int? Ea4 { get; set; } = 0;
 
-    public double? Ea5 { get; set; } = 0;
+    public int? Ea5 { get; set; } = 0;
 
-    public double? Ea6 { get; set; } = 0;
+    public int? Ea6 { get; set; } = 0;
 
     public int? Nrores { get; set; }
